requirements.txt
streamlit>=1.28.0
requests>=2.31.0
python-docx>=1.0.1
PyMuPDF>=1.23.7
SpeechRecognition>=3.10.0
